requirements.txt
alembic==1.8.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
flask-googlemaps==0.4.1
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
ghp-import==2.1.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
greenlet==1.1.2
htmlmin==0.1.12
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.3.0
mkdocs-material==5.5.14
mkdocs-material-extensions==1.0.3
mkdocs-minify-plugin==0.3.0
oauthlib==3.2.0
packaging==21.3
protobuf==4.21.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pymdown-extensions==9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.37
uritemplate==4.1.1
urllib3==1.26.9
watchdog==2.1.9
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
